+++ geopackage-ios/extension/elevation/GPKGGriddedTile.h
@@ -50,19 +50,9 @@ extern NSString * const GPKG_EGT_COLUMN_STANDARD_DEVIATION;
 @property (nonatomic, strong) NSDecimalNumber *scale;
 
 /**
- *  Insert the scale value as null as opposed to using the default value
- */
-@property (nonatomic) BOOL nullScale;
-
-/**
  * The offset to the 0 value
  */
 @property (nonatomic, strong) NSDecimalNumber *offset;
-
-/**
- *  Insert the offset value as null as opposed to using the default value
- */
-@property (nonatomic) BOOL nullOffset;
 
 /**
  *  Minimum value of this tile
@@ -96,13 +86,13 @@ extern NSString * const GPKG_EGT_COLUMN_STANDARD_DEVIATION;
  *
  *  @return scale as a multiple relative to the unit of measure
  */
--(double) getScaleOrDefault;
+-(double) scaleOrDefault;
 
 /**
  *  Get the offset or default value
  *
  *  @return offset to the 0 value
  */
--(double) getOffsetOrDefault;
+-(double) offsetOrDefault;
 
 @end